Modified to handle exception & Success message

+++ src/main/java/com/pruthvi/service/StudentDetailsApiService.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.pruthvi.dto.StudentInfo;
 import com.pruthvi.dto.StudentInfoResponse;
+import com.pruthvi.exception.ConcurrencyModificationException;
 import com.pruthvi.exception.ConstrainValidationException;
 import com.pruthvi.exception.InvalidInputException;
 import com.pruthvi.mapper.ServiceMapper;
@@ -43,16 +44,29 @@ public class StudentDetailsApiService {
 		if(validation.getOverallStatus().equals("Failure")) {
 			throw new ConstrainValidationException(validation.getDataItems());
 		}
+		/** In case StudentId passed as Null during Update take from Path Variable **/
+		if(null == requestPayload.getStudentID() && studentId!=0) {
+			requestPayload.setStudentID(studentId);
+		}
 		var studentEntity = mapper.mapStudentInfoToStudentDetailEntity(requestPayload);
-		studentRepo.save(studentEntity);
+		var response = studentRepo.save(studentEntity);
+		requestPayload.setStudentID(response.getStudentID());
+		studentResponse.setStatus("Success");
+		var message = studentId == 0 ? "Student Detail Added Successfully" : "Student Detail Updated Successfully";
+		studentResponse.setMessage(message);
 		studentResponse.setStudentDtlInfo(requestPayload);
 		return studentResponse;
 	}
 
-	public StudentInfoResponse deleteStudentDetail(Integer studentId) {
+	public StudentInfoResponse deleteStudentDetail(Integer studentId) throws ConcurrencyModificationException {
 		StudentDetailsApiService.log.debug("Entered Delete Student Details Method");
 		var studentResponse = new StudentInfoResponse();
+		if(Boolean.FALSE.equals(studentRepo.findById(studentId).isPresent())) {
+			throw new ConcurrencyModificationException("Data Lost due to Concurrent update , please Referesh the page");
+		}
 		studentRepo.deleteById(studentId);
+		studentResponse.setStatus("Success");
+		studentResponse.setMessage("Student Detail Deleted Successfully");
 		return studentResponse;
 	}
